Update README.md
The raw data is manipulated data to mimic a chocolate manufacturer/retailer. 
The first part of the project is to find the total cost of chocolate, if we source all the ingredients from the cheapest supplier and meet all the demand. 
The second part is to build a forecast model to predict the demand more accurately. It includes exploratory data analysis, feature selection, model selection, cross-validation and use permutation to test the new feture improtance in the selected model.  

Signed-off-by: ShuDataScience <107440813+ShuDataScience@users.noreply.github.com>